Improve discovery of maven descriptor embedded inside jar files
For maven packaging it is common to embed the coordinates inside the jar
file under the meta-inf director. Currently the SBOM generator can not
detect such jars as being possible maven artifacts.

This now enhances it in a way that after checking for the properties it
look inside the content if it can find a suitable descriptor and extract
the information from there.